// src/platform/packages/private/kbn-esql-editor/src/types.d.ts
interface ESQLVariableService {
    isCreateControlSuggestionEnabled: boolean;
    esqlVariables: ESQLControlVariable[];
    enableCreateControlSuggestion: () => void;
    disableCreateControlSuggestion: () => void;
    clearVariables: () => void;
    addVariable: (variable: ESQLControlVariable) => void;
}
export interface EsqlPluginStartBase {
    variablesService: ESQLVariableService;
    getLicense: () => Promise<ILicense | undefined>;
    isServerless: boolean;
    enrichSources: (sources: ESQLSourceResult[]) => Promise<ESQLSourceResult[]>;
}
export interface ESQLEditorDeps {
    core: CoreStart;
    data: DataPublicPluginStart;
    storage: Storage;
    uiActions: UiActionsStart;
    kql: KqlPluginStart;
    fieldsMetadata?: FieldsMetadataPublicStart;
    usageCollection?: UsageCollectionStart;
    cps?: CPSPluginStart;
    esql?: EsqlPluginStartBase;
}